requirements.txt

flask
line-bot-sdk
requests
python-dotenv
